requirements.txt
Django==1.7.4
colorama==0.3.1
dj-database-url==0.3.0
dj-static==0.0.6
django-cors-headers==1.0.0
django-toolbelt==0.0.1
djangorestframework==3.0.4
gunicorn==19.2.0
html5lib==0.999
requests==2.3.0
six==1.6.1
static3==0.5.1
